util/matrix: cut trailing / from SSO req, unbreak iOS Element
Element iOS will send a request to cause Synapse to redirect to the
SSO provider with a trailing slash:

    /_matrix/client/r0/login/sso/redirect/?...

This causes synapse to respond with an M_UNRECOGNIZED error. Thus, for
now, explicitly match on that URL and remove the trailing slash. nginx
takes care of the query parameters. This looks exactly like issue 4785
[1]; however this should have been fixed in all recent versions of
Element iOS.

[1]: https://github.com/vector-im/element-ios/issues/4785